+++ Hibernate-app/src/main/java/put/poznan/pl/michalxpz/hibernateapp/repository/CategoryRepository.java
@@ -6,4 +6,4 @@ import org.springframework.stereotype.Repository;
 import put.poznan.pl.michalxpz.hibernateapp.model.Category;
 
 @Repository
-public interface CategoryRepository extends JpaRepository<Category, Long> {}
+public interface CategoryRepository extends JpaRepository<Category, Integer> {}
