+++ Hibernate-app/src/main/java/put/poznan/pl/michalxpz/hibernateapp/repository/TagRepository.java
@@ -6,5 +6,5 @@ import org.springframework.stereotype.Repository;
 import put.poznan.pl.michalxpz.hibernateapp.model.Tag;
 
 @Repository
-public interface TagRepository extends JpaRepository<Tag, Long> {}
+public interface TagRepository extends JpaRepository<Tag, Integer> {}
